Add files via upload
Initial Capstone Project Milestone Report Draft

This was both an exercise to help in my data story skills as well as an outline of what data science skills I've used thus far and how I'd like to move forward with my capstone project for this course. Any input is welcome!